+++ 01_krepios/lang/de/de_player.lua
@@ -1,3 +1,5 @@
+local f = string.format
+
 My.Translator:register("de", {
     player_power_presets_label = "Voreinstellungen",
     player_power_presets_label_load = "Laden",
@@ -11,4 +13,11 @@ My.Translator:register("de", {
         " Plätzen zu speichern. Bitte beachten Sie, dass nur der aktuelle Zustand und nicht der eingestellte Zielzustend \z
         gespeichert wird."
     end,
+
+    player_mining_label = "Asteroid abbauen",
+    player_mining_error_invalid_target = "Es gibt keinen Asteroiden in Laserreichweite.",
+    player_mining_error_not_enough_power = function(minEnergy)
+        return f("Es werden mindestens %d Energie benötigt, um mit Abbau eines Asteroiden zu beginnen.", minEnergy)
+    end,
+    player_mining_error_invalid_station = "Das Beam Weapon System muss repariert, ausreichend gekühlt und vollständig mit Energie versorgt sein, damit ein Asteroid abgebaut werden kann.",
 })